requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psutil==5.9.5
plotly==5.17.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.1
torch==2.0.1
transformers==4.33.2
opencv-python==4.8.0.76
Pillow==10.0.1
python-dateutil==2.8.2
requests==2.31.0
python-dotenv==1.0.0
flask-login==0.6.2
flask-sqlalchemy==3.1.1
d3==0.0.2
tailwindcss==3.3.3
pytest==7.4.2
coverage==7.3.1
